fix(screamingface): share access credentials by audience
