+++ docs/work/2026-08-19-OME-899-authenticate-scoreboard-requests.md
@@ -1,9 +1,9 @@
 ---
 ticket: OME-899
 stack: screamingface
-status: done
+status: in_progress
 started: 2026-08-19
-finished: 2026-08-19
+finished:
 ---
 
 # OME-899 — Authenticate protected Scoreboard requests
@@ -18,6 +18,12 @@ reads interactive or risking duplicate score submissions.
 
 - `packages/screamingface/src/screamingface/client.py` — Scoreboard-origin authentication, replay
   declaration, and lifecycle ownership.
+- `packages/screamingface/src/screamingface/_access/` — service-neutral Access adapter and
+  audience-scoped in-memory credential sharing.
+- `packages/screamingface/src/screamingface/_scoreboard/leaderboards.py` — explicit domain-owned
+  replay declarations.
+- `packages/screamingface/tests/test_authentication.py` — same- and distinct-audience regression
+  coverage.
 - `packages/screamingface/tests/test_leaderboards.py` — sync and async protected-action regression
   coverage.
 - `docs/{tasks,spec,plan,work}/2026-08-19-OME-899-*` — required SDLC artifacts.
@@ -27,12 +33,16 @@ reads interactive or risking duplicate score submissions.
 - A protected score GET receives Scoreboard credentials and is declared replay-safe.
 - An idempotency-keyed score POST receives Scoreboard credentials and is declared replay-safe.
 - Engine credentials are not substituted for Scoreboard-origin credentials.
+- A Scoreboard origin proving the Engine audience reuses the existing credential without a second
+  browser login; a distinct audience starts a separate login.
 - Sync and async public interfaces return decoded `LeaderboardScore` values.
 - Focused tests, Ruff, Pyright, and the full ScreamingFace gate pass.
 
 ## Acceptance
 
 - `sf.leaderboards.submit(result.candidates[0])` needs no private header manipulation.
+- Calling it after one Engine login does not request a second browser login when the two origins
+  advertise the same Access audience.
 - Protected score reads authenticate through the same internal seam.
 - Public reads remain non-interactive when no Access challenge is returned.
 - Score submission replay remains idempotency-gated.
@@ -40,12 +50,14 @@ reads interactive or risking duplicate score submissions.
 
 ## Outcome
 
-- **Actual files:** sync/async Client Scoreboard-auth wiring and replay declaration;
-  Leaderboard regression/lifecycle tests; required OME-899 task, spec, plan, and work artifacts.
-- **Commits:** `4d27a2a6` — authenticate protected Scoreboard requests.
-- **Gates:** `python3 .claude/scripts/run_gates.py screamingface` — ALL GATES GREEN;
-  focused Client/Leaderboard suites — 73 passed; live protected `get_score` through the patched
-  public interface — expected `404 unknown_score`, no Access 302.
+- **Actual files:** service-neutral Access authentication with an audience-keyed credential store;
+  sync/async Client Scoreboard wiring and lifecycle ownership; explicit Leaderboards replay
+  declarations; same- and distinct-audience regressions; required OME-899 SDLC artifacts.
+- **Commits:** `4d27a2a6` — initial protected Scoreboard authentication; the current change refines
+  shared-audience credential reuse and lifecycle cleanup.
+- **Gates:** authentication suite — 73 passed; focused authentication/Leaderboard suites — 118
+  passed; the complete ScreamingFace gate is green, including Ruff, formatting, Pyright, coverage,
+  notebook validation, package build, and distribution checks.
 - **Deviations:** Cloudflare dev configuration changed from separate Engine/Scoreboard audiences
   to one multi-domain audience during diagnosis. The implementation remains origin-scoped so
   custom deployments with separate Access applications are also supported. No notebook changes
