+++ packages/screamingface/src/screamingface/_access/__init__.py
@@ -1,0 +1,3 @@
+"""Service-neutral Cloudflare Access authentication."""
+
+__all__: list[str] = []
